backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
setuptools==69.2.0
sqlparse==0.5.0
wheel==0.43.0
djangorestframework==3.15.1
django-cors-headers==4.1.0
PyJWT==2.7.0
psycopg2-binary==2.9.6
django-redis==5.3.0
django-cleanup==8.1.0
pillow==10.3.0
celery==5.4.0
django-celery-beat==2.6.0
